fix eager fetching in locators
